Update editor text as color is changed
Make color-picker even more awesome when used in combination with an
as-you-type CSS reloader. In this scenario, a developer can keep the
color picker open as they see the impact of their color choice in
the reloaded interface.

As a bonus, the undo history can walk backwards through color selections.